Move database queries to ferrum-db crate
