+++ ferrum-db/Cargo.toml
@@ -12,6 +12,8 @@ serde = "1.0"
 chrono = { version = "0.4", features = ["serde"] }
 uuid = { version = "0.8", features = ["v4", "serde"] }
 argon2 = { version = "0.3", features = ["std"] }
+anyhow = "1.0.40"
+tracing = "0.1.19"
 
 [dev-dependencies]
 fake = "~2.3.0"
