+++ ferrum-db/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod users;
-pub mod servers;
 pub mod channels;
 pub mod messages;
+pub mod servers;
+pub mod users;
