+++ src/routes/channels/create_message.rs
@@ -5,22 +5,16 @@ use actix_http::StatusCode;
 use actix_web::{web, HttpResponse, ResponseError};
 use anyhow::Context;
 use ferrum_db::{
-    messages::models::{MessageContent, MessageModel, NewMessage},
-    users::models::UserModel,
+    messages::{
+        models::{MessageContent, NewMessage},
+        queries::insert_message,
+    },
+    users::queries::does_user_have_access_to_channel,
 };
-use sqlx::{PgPool, Postgres, Transaction};
+use sqlx::PgPool;
 use uuid::Uuid;
 
-use crate::{
-    domain::messages::MessageResponse,
-    error_chain_fmt,
-    jwt::AuthorizationService,
-    utilities::does_user_have_access_to_channel,
-    websocket::{
-        messages::{SendMessageToChannel, WebSocketMessage},
-        Server,
-    },
-};
+use crate::{error_chain_fmt, jwt::AuthorizationService, websocket::Server};
 
 #[derive(serde::Deserialize)]
 pub struct BodyData {
@@ -63,12 +57,12 @@ impl ResponseError for CreateMessageError {
     }
 }
 
-#[tracing::instrument(name = "Create a new channel message", skip(body, pool, auth, server), fields(user_id = %auth.claims.id, user_email = %auth.claims.email))]
+#[tracing::instrument(name = "Create a new channel message", skip(body, pool, auth, _server), fields(user_id = %auth.claims.id, user_email = %auth.claims.email))]
 pub async fn create_message(
     channel_id: web::Path<Uuid>,
     body: web::Json<BodyData>,
     pool: web::Data<PgPool>,
-    server: web::Data<Addr<Server>>,
+    _server: web::Data<Addr<Server>>,
     auth: AuthorizationService,
 ) -> Result<HttpResponse, CreateMessageError> {
     let new_message: NewMessage = body
@@ -85,7 +79,7 @@ pub async fn create_message(
         .await
         .context("Failed to acquire a postgres connection from pool.")?;
 
-    let message = insert_message(
+    let _message = insert_message(
         &mut transaction,
         &pool,
         &new_message,
@@ -95,58 +89,17 @@ pub async fn create_message(
     .await
     .context("Failed to insert a new channel message to the database.")?;
 
-    transaction
-        .commit()
-        .await
-        .context("Failed to commit SQL transaction to store a new channel message.")?;
+    todo!("Convert MessageModel to MessageResponse");
 
-    server.do_send(SendMessageToChannel::new(
-        *channel_id,
-        WebSocketMessage::NewMessage { message },
-    ));
+    // transaction
+    //     .commit()
+    //     .await
+    //     .context("Failed to commit SQL transaction to store a new channel message.")?;
 
-    Ok(HttpResponse::Ok().finish())
-}
+    // server.do_send(SendMessageToChannel::new(
+    //     *channel_id,
+    //     WebSocketMessage::NewMessage { message },
+    // ));
 
-#[tracing::instrument(
-    name = "Saving a new channel message to the database",
-    skip(transaction, new_message, channel_id, user_id)
-)]
-async fn insert_message(
-    transaction: &mut Transaction<'_, Postgres>,
-    pool: &PgPool,
-    new_message: &NewMessage,
-    channel_id: Uuid,
-    user_id: Uuid,
-) -> Result<MessageResponse, sqlx::Error> {
-    let id = Uuid::new_v4();
-
-    let message = sqlx::query_as!(
-        MessageModel,
-        r#"
-        INSERT INTO messages (id, channel_id, user_id, content)
-        VALUES ($1, $2, $3, $4)
-        RETURNING *
-        "#,
-        id,
-        channel_id,
-        user_id,
-        new_message.content.as_ref(),
-    )
-    .fetch_one(transaction)
-    .await?;
-
-    let user = sqlx::query_as!(
-        UserModel,
-        r#"
-        SELECT id, username, email, created_at, updated_at
-        FROM users
-        WHERE users.id = $1
-        "#,
-        message.user_id,
-    )
-    .fetch_one(pool)
-    .await?;
-
-    Ok(MessageResponse::new(&message, &user))
+    // Ok(HttpResponse::Ok().finish())
 }
