+++ src/routes/servers/create.rs
@@ -4,9 +4,11 @@ use actix::Addr;
 use actix_http::StatusCode;
 use actix_web::{web, HttpResponse, ResponseError};
 use anyhow::Context;
-use ferrum_db::servers::models::{NewServer, ServerModel, ServerName};
-use sqlx::{PgPool, Postgres, Transaction};
-use uuid::Uuid;
+use ferrum_db::servers::{
+    models::{NewServer, ServerName},
+    queries::{add_default_channel_to_server, add_user_to_server, insert_server},
+};
+use sqlx::PgPool;
 
 use crate::{
     error_chain_fmt,
@@ -86,77 +88,4 @@ pub async fn create(
     websocket_server.do_send(messages::NewServer::new(auth.claims.id, server.id));
 
     Ok(HttpResponse::Ok().finish())
-}
-
-#[tracing::instrument(
-    name = "Saving a new server to the database",
-    skip(transaction, new_server)
-)]
-async fn insert_server(
-    transaction: &mut Transaction<'_, Postgres>,
-    new_server: &NewServer,
-    user_id: Uuid,
-) -> Result<ServerModel, sqlx::Error> {
-    let server = sqlx::query_as!(
-        ServerModel,
-        r#"
-        INSERT INTO servers (id, name, owner_id)
-        VALUES ($1, $2, $3)
-        RETURNING *
-        "#,
-        Uuid::new_v4(),
-        new_server.name.as_ref(),
-        user_id,
-    )
-    .fetch_one(transaction)
-    .await?;
-
-    Ok(server)
-}
-
-#[tracing::instrument(
-    name = "Saving a new default server channel to the database",
-    skip(transaction, server_id)
-)]
-async fn add_default_channel_to_server(
-    transaction: &mut Transaction<'_, Postgres>,
-    server_id: Uuid,
-) -> Result<(), sqlx::Error> {
-    sqlx::query!(
-        r#"
-        INSERT INTO channels (id, server_id, name) VALUES ($1, $2, $3)
-        "#,
-        Uuid::new_v4(),
-        server_id,
-        "general",
-    )
-    .execute(transaction)
-    .await?;
-
-    Ok(())
-}
-
-#[tracing::instrument(
-    name = "Saving a new users_servers entry to the database",
-    skip(transaction, user_id, server_id)
-)]
-async fn add_user_to_server(
-    transaction: &mut Transaction<'_, Postgres>,
-    user_id: Uuid,
-    server_id: Uuid,
-) -> Result<(), sqlx::Error> {
-    let id = Uuid::new_v4();
-
-    sqlx::query!(
-        r#"
-        INSERT INTO users_servers (id, user_id, server_id) VALUES ($1, $2, $3)
-        "#,
-        id,
-        user_id,
-        server_id,
-    )
-    .execute(transaction)
-    .await?;
-
-    Ok(())
 }
